fix(evals): close four ways a bad API key placement scored green
The client env scan skipped `export VITE_ADMIN=sb_secret_...`, because the
`export ` prefix broke the name match, and skipped a secret behind a prefix
declared through a variable, because only a literal `envPrefix` parsed. It now
reads every `.env` outside `supabase/` and fails on a secret in any of them,
parsing neither names nor prefixes. A secret in the client's env is exposed
whichever name holds it.

The server check matched variable names only, so a legacy key bound to an
alias passed. It now also decodes JWTs under `supabase/functions` and fails on
an `anon` or `service_role` role. Renamed to `no legacy key reaches the
server`, which is what it now proves.

`client source calls signUp` overstated a literal match, since dead code
satisfies it. Renamed to `client source contains a signUp call`. Proving the
screen works needs a driven DOM the scorer does not have, so the claim is
narrowed rather than strengthened.

Credential detection moves to keys.ts, now that the bundle scan and the server
scan both need it.
